refactor: moved dynamic layout controls to common components
